Fix a nasty, insidious bug in editorReadKey
The condition in the while was wrong, which meant it would break out of
the loop whenever `read()` actually read a character _or_ timed out.  If
it times out, it does nothing to the char c, so the uninitialized (!)
value gets returned.  Eventually we got into a situation where the same
bit of memory would be reused the next time editorReadKey was called.
So if you hit a cursor key to move the cursor, editorReadKey would
return that value again and again every call unless you hit another key.
This also caused weird things with key sequences.  In order to process a
key sequence, there has to be a bit of state, and if the function was in
the wrong state when the key was pressed, it would get misread.